update check fir service env support
