+++ libs/python/helperEnvKyma.py
@@ -1,15 +1,17 @@
+import logging
+import os
+import sys
+import time
+
+from libs.python.helperCommandExecution import (runShellCommand,
+                                                runShellCommandFlex)
+from libs.python.helperEnvironments import \
+    check_if_service_plan_supported_in_environment
 from libs.python.helperGeneric import getTimingsForStatusRequest
 from libs.python.helperJson import convertStringToJson
 from libs.python.helperYaml import (
     build_and_store_service_binding_yaml_from_parameters,
-    build_and_store_service_instance_yaml_from_parameters,
-)
-from libs.python.helperCommandExecution import runShellCommand, runShellCommandFlex
-import os
-import sys
-import logging
-import time
-from libs.python.helperServiceInstances import check_if_service_plan_supported_in_environment
+    build_and_store_service_instance_yaml_from_parameters)
 
 log = logging.getLogger(__name__)
